Air_New_V1.1 first commit.
Air_New_V1.1
It is a system developed for a company that converts gasoline vehicles to LP gas.

This system is client-server type. The system allows you to connect to a SQL Server database, to register new clients, it also allows you to consult registered clients, these records can be exported in an Excel file. It also includes a calculation module to calculate certain variables of value for said company.
It has an email and password registration, and then allows authentication through saved credentials.